+++ packages/generator/src/utils/get-blitz-dependency-version.ts
@@ -6,9 +6,10 @@ export const getBlitzDependencyVersion = async (
   cliVersion: string,
 ): Promise<Fallbackable<string>> => {
   try {
+    // TODO: Need to update this to handle alpha, beta and major
     const {alpha} = await fetchDistTags("blitz")
 
-    if (cliVersion.includes("alpha")) {
+    if (alpha) {
       return {value: alpha}
     }