+++ packages/config/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @blitzjs/config
 
+## 2.0.0-alpha.13
+
 ## 2.0.0-alpha.12
 
 ## 2.0.0-alpha.11