+++ packages/generator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @blitzjs/generator
 
+## 2.0.0-alpha.13
+
+### Patch Changes
+
+- Fix codegen and postinstall to make work with pnpm
+
 ## 2.0.0-alpha.12
 
 ### Patch Changes